plantear cambiar dotenv

+++ routes/user.routes.mjs
@@ -3,11 +3,15 @@ import bcrypt from 'bcrypt'
 import { Image } from '../models/image.model.mjs'
 import { fileUploader } from '../config/cloudinary.config.mjs'
 import { validateImage, validatePartialImage } from '../schemas/image.mjs'
+import { validateCode } from '../schemas/code.mjs'
+import dotenv from 'dotenv'
+dotenv.config()
 
 const router = express.Router()
 
-const saltRounds = 10
+const USER_CODE = process.env.USER_CODE
 
+console.log('user', USER_CODE)
 router.get('/', (req, res, next) => {
   // Pensar si vale la pena este filtro
   /*   const { label } = req.query
@@ -57,6 +61,12 @@ router.patch('/:id', (req, res, next) => {
 
 router.delete('/:id', (req, res, next) => {
   const { id } = req.params
+  const result = validateCode(req.body)
+  console.log(USER_CODE, result.data.userCode)
+  if (!bcrypt.compareSync(result.data.userCode, USER_CODE)) {
+    return res.status(403).json({ error: 'Your user code is not correct !' })
+  }
+
   Image.findByIdAndDelete(id)
     .then((response) => {
       res.status(204).json({ response: 'ok' })
